feat(acp_zed): register LiteAgent as a native agent type
`LiteAgent` joins `RequirementAgent`, `ToolCallingAgent`, and `ReActAgent` as a
pre-registered factory. It streams through an emitter `final_answer` callback
rather than an async iterator, so the factory bridges the synchronous callback
to the async ACP connection via an `asyncio.Queue`. After the turn settles,
tool-call entries and tool results accumulated in memory are drained through
`_stream_messages_since` so they surface in Zed's agent panel.

The `examples/serve/acp_zed_lite.py` example drops its inline factory
registration — users now just `.register(LiteAgent(...)).serve()`.
